+++ lib/command/run.js
@@ -2,7 +2,6 @@
 let getConfig = require('./utils').getConfig;
 let getTestRoot = require('./utils').getTestRoot;
 let deepMerge = require('./utils').deepMerge;
-let replaceValue = require('./utils').replaceValue;
 let Codecept = require('../codecept');
 var fork = require('child_process').fork
 let output = require('../output');
@@ -131,4 +130,26 @@ module.exports = function (suite, test, options) {
 
     }
   }
+
+  function replaceValue(obj, key, value) {
+    /* search key in object recursive and replace value in it
+     */
+    if (obj instanceof Array) {
+      for (var i in obj) {
+        replaceValue(obj[i], key, value);
+      }
+    }
+    if (obj[key]) obj[key] = value;
+
+
+    if ((typeof obj == "object") && (obj !== null)) {
+      var children = Object.keys(obj);
+      if (children.length > 0) {
+        for (i = 0; i < children.length; i++) {
+          replaceValue(obj[children[i]], key, value);
+        }
+      }
+    }
+    return obj;
+  };
 };
