+++ lib/utils.js
@@ -122,25 +122,3 @@ module.exports.test = {
     throw new Error('should not be thrown');
   }
 };
-
-module.exports.replaceValue = function(obj, key, value) {
-  /* search key in object recursive and replace value in it
-   */
-  if (obj instanceof Array) {
-    for (var i in obj) {
-      replaceValuesHelper(obj[i], key, value);
-    }
-  }
-  if (obj[key]) obj[key] = value;
-
-
-  if ((typeof obj == "object") && (obj !== null)) {
-    var children = Object.keys(obj);
-    if (children.length > 0) {
-      for (i = 0; i < children.length; i++) {
-        replaceValuesHelper(obj[children[i]], key, value);
-      }
-    }
-  }
-  return obj;
-};
